requirements.txt
django
djangorestframework
scikit-learn